lua: add example how to get events for mouse button down + mouse move
The ancient and forgotten support for multi-key bindings and the fact
that mouse move events are tracked the same as key events makes this
magically work. But it's certainly not perfect.

This will show a "-" as mouse cursor as long as testl.lua thinks the
mouse is down + moved, and "+" otherwise. There's no explicit way to
know whether/when the mouse button was released.