osc: Replaced all show/hide mechanics with deadzone concept
It now entirely depends on if the mouse is in the deadzone or not if the OSC is shown or not. The OSC will show up on leaving the deadzone and will hide immedietly on entering it.
The deadzone always spans over the oposite part the screen from the OSC. The distance between deadzone and the OSC can be configured with deadzonedist.