osc: replace deadzone handling with mouse input area
set_mouse_area() sets the region where mouse events should be received.
Outside this area, mouse events are handled normally (and for example
use the user's normal input.conf bindings). Also, leaving the mouse area
triggers the mouse_leave binding. (Mouse enter is already covered by
mouse_move.)

So, set the mouse area to the OSC (non-deadzone) area. Make sure the OSC
is visible when entering the area (mouse_move), and hiding the OSC when
the mouse goes outside (mouse_leave). As a bonus, it will be possible to
hide the OSC if the mouse cursor leaves the window entirely (but that
requires support from the VO).